Change relative zone file path to absolute path
When using a zone file path that isn't the default `/etc/bind`, `addzone` fails because the RNDC call uses a relative path, resulting in RNDC failing because it can't find the new zone file in the `/etc/bind` directory. By referencing the new zone file with it's absolute path, we guarantee that regardless of where the zone files are stored, RNDC will correctly be able to find it.